+++ CHANGELOG.md
@@ -10,9 +10,7 @@ The structure and content of this file follows [Keep a Changelog](https://keepac
 - "exists" is now an alias for the "has" filter operation.
 - Added length, count, match, and search functions.
 - Added `Nothing` as a value for comparison to return values where nothing is found.
-
-- TBD support [?foo
-
+- Added support no parenthesis around a filter so `[?@.x == 3]` is now valid.
 
 ## [1.17.5] - 2023-02-19
 ### Added